CMake edit.
