+++ Project/classes/header/dummy.h
@@ -1,0 +1,11 @@
+#pragma once
+#include <iostream>
+
+
+class Dummy{
+private:
+    int x, y;
+public:
+    Dummy(int x_, int y_);
+    static int test();
+};
